core/api/serializers/recommendation_serializer.py
from rest_framework.serializers import ModelSerializer
from core.models import *

class RecommendationSerializer(ModelSerializer):

    class Meta:
        model = Recommendation
        fields = '__all__'